refactor(secondary-screen): promote app dock to a widget
Extract the Now Playing app dock render (`_buildAppDock` + `_buildDockSlot`)
out of `_SecondaryScreenState` into a new `const`-constructor `AppDock` widget
under `secondary_screen/widgets/`. The dock is a pure, input-driven subtree:
the host passes the current state snapshot plus the three slot callbacks
(launch / pick-slot / clear-slot), so the widget re-reads no provider state of
its own — identical render and behavior.

The shared `_buildDockIcon` helper (used by both dock slots and the app-picker
tiles) is promoted to a top-level `buildDockIcon` function alongside the
widget, so both call sites share one copy. Dropped the now-unnecessary
`flutter/services.dart` host import (its only use, `Uint8List`, moved with the
icon helper).

Host 2209 -> ~2130 lines. Bodies moved verbatim; no behavior change.
